Abdul Ghaffar & Sons overseas employment  agencies in Pakistan
"Abdul Ghaffar & Sons" | overseas employment  agencies in Pakistan is one of the leading manpower recruitment agencies in Pakistan, supplying manpower to entire Middle Eastern region especially in UAE and Saudi Arabia. We are providing the best recruitment options and elite employment service in Pakistan since 1986. The company was established with the ultimate vision of serving Pakistan with most reliable overseas recruitment services and tops the list of manpower recruitment agencies in Pakistan. In the past three decades, we have set great records of recruiting several thousands of workers in Middle East with best employment options and manpower benefits.

"Abdul Ghaffar & Sons" is the most renowned overseas employment agency in Pakistan, which is also authentic and licensed recruitment agency. We hold a license approved by Ministry of Labor, manpower and overseas Pakistanis, Government of Pakistan. Our overseas employment agency license no